update ShellFlyComponent updateForward function

+++ Project/3D_Tank/3D_Tank/ShellFlyComponent.h
@@ -13,7 +13,7 @@ public:
 private:
 	GameObject* target;
 	Vector3 finalForward;
-	int rotateSpeed;
+	float rotateSpeed;
 	float angle;
 	//float angle;
